requirements.txt
Django==1.8
dj-database-url==0.3.0
django-pipeline==1.4.7
whitenoise==1.0.6
https://github.com/jezdez/django-configurations/archive/5ece1070448d3e026b81422691c60207f5699c0c.zip#egg=django-configurations==0.8
djangorestframework==3.1.1
